create-challenge코드 수정
